+++ docs/src/components/mdxComponents/MDLink.tsx
@@ -1,6 +1,7 @@
 import React, { ReactElement } from "react";
 import { styled } from "linaria/react";
 import Link from "next/link";
+import { useRouter } from "next/router";
 import { media, tm, tmDark, tmSelectors } from "../../themes";
 import ExternalLinkIcon from "../../assets/icons/external-link-icon";
 
@@ -56,14 +57,27 @@ const StyledMdLinkContainer = styled.span`
   }
 `;
 
-const getPathFromHref = (href: string) => {
-  const pathname = href
+const getAbsoluteHrefFromRelativePath = (href: string, currentHref: string) => {
+  const pathSegments = currentHref
     .split("/")
-    .filter((hrefPart: string) => ![".", ".."].includes(hrefPart))
-    .join("/")
-    .toLowerCase();
+    .filter((segment) => segment !== "");
+  const hrefSegments = href.split("/").filter((segment) => segment !== ".");
 
-  return pathname.startsWith("/") ? pathname : `/${pathname}`;
+  const pathSegmentsCount = pathSegments.length;
+  const upperLevelsCount = hrefSegments.filter(
+    (segment) => segment === ".."
+  ).length;
+
+  const baseSegmentsCount = Math.max(
+    pathSegmentsCount - 1 - upperLevelsCount,
+    0
+  );
+
+  const baseSegments = pathSegments.slice(0, baseSegmentsCount + 1);
+
+  const newSegments = ["", ...baseSegments, ...hrefSegments];
+
+  return newSegments.join("/").replace("/index", "");
 };
 
 const renderLinkByType = ({
@@ -71,7 +85,14 @@ const renderLinkByType = ({
   href,
   isExternalLink,
   isAnchor,
-}: Props & { isExternalLink: boolean; isAnchor: boolean }) => {
+  isAbsoluteLink,
+  currentHref,
+}: Props & {
+  isExternalLink: boolean;
+  isAnchor: boolean;
+  isAbsoluteLink: boolean;
+  currentHref: string;
+}) => {
   if (isExternalLink) {
     return (
       <a href={href} target="_blank" rel="noreferrer">
@@ -83,9 +104,17 @@ const renderLinkByType = ({
   if (isAnchor) {
     return <a href={href}>{children}</a>;
   }
+  if (isAbsoluteLink) {
+    return (
+      <Link href={href}>
+        {/* eslint-disable-next-line */}
+        <a>{children}</a>
+      </Link>
+    );
+  }
 
   return (
-    <Link href={getPathFromHref(href.replace(/\.md$/, ""))}>
+    <Link href={getAbsoluteHrefFromRelativePath(href, currentHref)}>
       {/* eslint-disable-next-line */}
       <a>{children}</a>
     </Link>
@@ -93,12 +122,21 @@ const renderLinkByType = ({
 };
 
 const MDLink = ({ children, href }: Props) => {
+  const router = useRouter();
   const isExternalLink = href.startsWith("http");
+  const isAbsoluteLink = href.startsWith("/");
   const isAnchor = href.startsWith("#");
 
   return (
     <StyledMdLinkContainer>
-      {renderLinkByType({ href, children, isAnchor, isExternalLink })}
+      {renderLinkByType({
+        href: href.replace(/\.mdx?$/, ""),
+        children,
+        isAnchor,
+        isExternalLink,
+        isAbsoluteLink,
+        currentHref: router.asPath,
+      })}
     </StyledMdLinkContainer>
   );
 };