+++ docs/src/components/ui/DesktopMenu.tsx
@@ -347,7 +347,7 @@ const DesktopMenu = ({
               <SocialLink
                 target="_blank"
                 rel="noreferrer"
-                href={social.name}
+                href={social.href}
                 aria-label={social.name}
               >
                 <Icon />