+++ docs/src/global-tabs.tsx
@@ -4,7 +4,7 @@ type TabType = string;
 
 interface ISelectedTabsState {
   "npm/yarn": string;
-  "js/ts": string;
+  "TypeScript/JavaScript": string;
   [key: TabType]: string;
 }
 
@@ -16,13 +16,13 @@ interface ITabsContext {
 export const GlobalTabsContext = React.createContext<ITabsContext>({
   tabsState: {
     "npm/yarn": "npm",
-    "js/ts": "js",
+    "TypeScript/JavaScript": "TypeScript",
   },
   changeTab: () => {},
 });
 
 export const generateTabsGroupType = (options: string): string => {
-  return options.split(",").sort().join("/");
+  return options.split(",").join("/");
 };
 
 export const TabsProvider = ({
@@ -30,7 +30,7 @@ export const TabsProvider = ({
 }: React.PropsWithChildren<{}>): JSX.Element => {
   const [tabsState, setTabsState] = useState<ISelectedTabsState>({
     "npm/yarn": "npm",
-    "js/ts": "js",
+    "TypeScript/JavaScript": "TypeScript",
   });
 
   const changeTab = useCallback(