+++ packages/hardhat-core/src/internal/core/jsonrpc/types/input/hardhat-network.ts
@@ -8,6 +8,7 @@ export const rpcForkConfig = optional(
     {
       jsonRpcUrl: t.string,
       blockNumber: optional(t.number),
+      httpHeaders: optional(t.record(t.string, t.string, "httpHeaders")),
     },
     "RpcForkConfig"
   )