+++ scripts/check-dependencies.js
@@ -3,10 +3,10 @@ const path = require("path");
 
 // An array of dependencies whose version checks are ignored for all the
 // packages
-const IGNORE_FROM_ALL = ["web3", "hardhat"];
+const IGNORE_SAME_VERSION_FROM_ALL = ["web3", "hardhat"];
 
 // A map from dependencies to package names where it should be ignored
-const IGNORE_FOR_PACKAGES = {
+const IGNORE_SAME_VERSION_FOR_PACKAGES = {
   chai: ["@nomiclabs/hardhat-truffle4", "@nomiclabs/hardhat-truffle5"],
   "@types/chai": ["@nomiclabs/hardhat-truffle4", "@nomiclabs/hardhat-truffle5"],
   "truffle-contract": [
@@ -20,10 +20,15 @@ const IGNORE_FOR_PACKAGES = {
   eslint: ["docs"],
   ["eslint-config-prettier"]: ["docs"],
   ["eslint-plugin-prettier"]: ["docs"],
-  ["typescript"]: ["docs"],
   ["glob"]: ["docs"],
   ["undici"]: ["docs"],
-  ["ts-node"]: ["docs"],
+  ["ts-node"]: ["docs", "hardhat"],
+  ["typescript"]: ["docs", "hardhat"],
+};
+
+const IGNORE_PEER_DEPENDENCIES_CHECK_FOR_PACKAGES = {
+  typescript: ["hardhat"],
+  ["ts-node"]: ["hardhat"],
 };
 
 function checkPeerDepedencies(packageJson) {
@@ -41,6 +46,14 @@ function checkPeerDepedencies(packageJson) {
 
   let success = true;
   for (const dependency of Object.keys(packageJson.peerDependencies)) {
+    if (
+      IGNORE_PEER_DEPENDENCIES_CHECK_FOR_PACKAGES[dependency]?.includes(
+        packageJson.name
+      )
+    ) {
+      continue;
+    }
+
     if (packageJson.devDependencies[dependency] === undefined) {
       console.error(
         `${packageJson.name} has ${dependency} as peerDependency, but not as devDependency`
@@ -72,13 +85,13 @@ function addDependencies(packageName, dependenciesToAdd, allDependenciesMap) {
   }
 
   for (const [name, spec] of Object.entries(dependenciesToAdd)) {
-    if (IGNORE_FROM_ALL.includes(name)) {
+    if (IGNORE_SAME_VERSION_FROM_ALL.includes(name)) {
       continue;
     }
 
     if (
-      IGNORE_FOR_PACKAGES[name] !== undefined &&
-      IGNORE_FOR_PACKAGES[name].includes(packageName)
+      IGNORE_SAME_VERSION_FOR_PACKAGES[name] !== undefined &&
+      IGNORE_SAME_VERSION_FOR_PACKAGES[name].includes(packageName)
     ) {
       continue;
     }