+++ packages/hardhat-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # hardhat
 
+## 2.9.9
+
+### Patch Changes
+
+- ce6136b2a: Fix incorrect peerDependency
+- 3fd3756e6: Specify ts-node and typescript as optional peerDependencies
+
+## 2.9.8
+
+### Patch Changes
+
+- 5bd775a8b: Added support for HTTP headers in the forking config (thanks @TimDaub!)
+- 3770a5bbc: Expose a missing method in the Artifacts interface
+
 ## 2.9.7
 
 ### Patch Changes